main.py
import time
import json
from dotenv import load_dotenv
import os

# Load Environment
load_dotenv()

def main():
    print("🚀 Initializing AlgoTrade Fleet...")
    
    api_key = os.getenv("BINANCE_API_KEY")
    if not api_key or "your_api_key" in api_key:
        print("❌ Error: .env file not configured.")
        return

    print("✅ Environment loaded.")
    print("📡 Connecting to Binance Futures...")
    
    # TODO: Initialize Engine Here
    # engine = TradingEngine()
    # engine.start()

    while True:
        print("💓 System Heartbeat - Waiting for strategy execution...")
        time.sleep(10)

if __name__ == "__main__":
    main()
