requirements.txt
ccxt
pandas
pandas_ta
python-dotenv
fastapi
uvicorn
pydantic
requests
